The StructureDiagramGenerator now generates coordinates for single condensed ringsystems. The underlying method RingPlacer.placeRing() should also be used for adding rings in an interactive mode.
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@59 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ org/openscience/cdk/Bond.java
@@ -239,7 +239,7 @@ public class Bond extends ElectronContainer implements Cloneable
 			x += getAtomAt(f).getX2D();
 			y += getAtomAt(f).getY2D();
 		}
-		return new Point2d(x / getAtomCount(), y / getAtomCount());
+		return new Point2d(x / ((double)getAtomCount()), y / ((double)getAtomCount()));
 	}
